# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Category {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  slug: string;
  description?: string;
  descriptionAr?: string;
  descriptionFr?: string;
  imageUrl?: string;
  icon?: string;
  productCount?: number;
  createdAt?: string;
}

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  slug: string;
  description?: string;
  descriptionAr?: string;
  descriptionFr?: string;
  price: number;
  originalPrice?: number | null;
  categoryId: number;
  categoryName: string;
  categoryNameAr: string;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  stock: number;
  isFeatured: boolean;
  isNew: boolean;
  badge?: string | null;
  createdAt?: string;
}

export type ProductDetailSpecifications = { [key: string]: string };

export type ProductDetail = Product & {
  specifications?: ProductDetailSpecifications;
  tags?: string[];
};

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface OrderItem {
  productId: number;
  quantity: number;
  price: number;
}

export interface CreateOrderRequest {
  fullName: string;
  phone: string;
  regionId: number;
  cityId?: number | null;
  address?: string | null;
  items: OrderItem[];
  isInstant?: boolean;
  notes?: string | null;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  fullName: string;
  phone: string;
  regionId: number;
  regionName: string;
  address?: string | null;
  total: number;
  items: OrderItem[];
  createdAt: string;
}

export interface FavoriteRequest {
  productId: number;
  sessionId: string;
}

export interface Region {
  id: number;
  code: string;
  name: string;
  nameAr: string;
  nameFr: string;
  deliveryDays: number;
  deliveryFee: number;
}

export interface City {
  id: number;
  name: string;
  nameAr: string;
  regionId: number;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  sortBy?: ListProductsSortBy;
  page?: number;
  limit?: number;
  featured?: boolean;
};

export type ListProductsSortBy =
  (typeof ListProductsSortBy)[keyof typeof ListProductsSortBy];

export const ListProductsSortBy = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
  popular: "popular",
} as const;

export type GetFavoritesParams = {
  sessionId: string;
};

export type RemoveFavoriteParams = {
  sessionId: string;
};