# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QuickDz Algerian Marketplace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Product categories
  - name: products
    description: Products
  - name: cart
    description: Shopping cart
  - name: orders
    description: Orders
  - name: favorites
    description: Favorites
  - name: regions
    description: Algerian regions (wilayas)
  - name: auth
    description: Authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Products list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (instant or full checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get favorites list
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add product to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteRequest"
      responses:
        "201":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /favorites/{productId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove product from favorites
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed from favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /regions:
    get:
      operationId: listRegions
      tags: [regions]
      summary: List all Algerian wilayas
      responses:
        "200":
          description: List of wilayas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"

  /regions/{id}/cities:
    get:
      operationId: getCitiesByRegion
      tags: [regions]
      summary: Get cities by wilaya ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        slug:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        descriptionFr:
          type: string
        imageUrl:
          type: string
        icon:
          type: string
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - nameAr
        - nameFr
        - slug

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        slug:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        descriptionFr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryNameAr:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        stock:
          type: integer
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        badge:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - nameAr
        - nameFr
        - slug
        - price
        - categoryId
        - categoryName
        - categoryNameAr
        - imageUrl
        - rating
        - reviewCount
        - stock
        - isFeatured
        - isNew

    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            specifications:
              type: object
              additionalProperties:
                type: string
            tags:
              type: array
              items:
                type: string

    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages

    CreateOrderRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        regionId:
          type: integer
        cityId:
          type: integer
          nullable: true
        address:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        isInstant:
          type: boolean
        notes:
          type: string
          nullable: true
      required:
        - fullName
        - phone
        - regionId
        - items

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - quantity
        - price

    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        fullName:
          type: string
        phone:
          type: string
        regionId:
          type: integer
        regionName:
          type: string
        address:
          type: string
          nullable: true
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - status
        - fullName
        - phone
        - regionId
        - regionName
        - total
        - items
        - createdAt

    FavoriteRequest:
      type: object
      properties:
        productId:
          type: integer
        sessionId:
          type: string
      required:
        - productId
        - sessionId

    Region:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        deliveryDays:
          type: integer
        deliveryFee:
          type: number
      required:
        - id
        - code
        - name
        - nameAr
        - nameFr
        - deliveryDays
        - deliveryFee

    City:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        regionId:
          type: integer
      required:
        - id
        - name
        - nameAr
        - regionId