# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface CreateOrderRequest {
  fullName: string;
  phone: string;
  regionId: number;
  cityId?: number | null;
  address?: string | null;
  items: OrderItem[];
  isInstant?: boolean;
  notes?: string | null;
}