# lib/api-zod/src/generated/types/favoriteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteRequest {
  productId: number;
  sessionId: string;
}