# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  slug: string;
  description?: string;
  descriptionAr?: string;
  descriptionFr?: string;
  price: number;
  originalPrice?: number | null;
  categoryId: number;
  categoryName: string;
  categoryNameAr: string;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  stock: number;
  isFeatured: boolean;
  isNew: boolean;
  badge?: string | null;
  createdAt?: Date;
}